#1 Create Component, Document and Style interfaces with related classes
This update introduces multiple new interfaces and related classes for the Component, Document and Style of the kpdf API. It also modifies BoxModel from a data class to an interface with a Box class implementation. An enum class was added for ComponentType. Altogether, these changes enhance the structure and flexibility of the API.